move enums to env file
